Added functionality to change Text Size across App
Created a global font size so that the user can dynamically change the font size across the app. Within the settings page there is now a text size slider for this functionality. I had to import a react package into the .json file for the sider component, so you might need to run the same command in the terminal: npm install @react-native-community/slider